Compile fix 2: CPUAccessFlags is the enum, not a uint
CS0266 at Canvas.cs(1708,34). CsWin32 projects D3D11_TEXTURE2D_DESC.CPUAccessFlags
as D3D11_CPU_ACCESS_FLAG, so the cast to uint I wrote from the C header's shape
is the wrong direction. The neighbouring descs assign a literal 0, which is legal
for an enum and is why nothing here read as a type.

The WMC9999 "Object reference not set" from the XAML markup compiler in the same
log is downstream of this error, not a second defect.

+++ prototypes/sb_winui/Canvas.cs
@@ -1705,7 +1705,7 @@ internal sealed unsafe class Canvas : IDisposable
                 SampleDesc = new DXGI_SAMPLE_DESC { Count = 1, Quality = 0 },
                 Usage = D3D11_USAGE.D3D11_USAGE_STAGING,
                 BindFlags = 0,
-                CPUAccessFlags = (uint)D3D11_CPU_ACCESS_FLAG.D3D11_CPU_ACCESS_READ,
+                CPUAccessFlags = D3D11_CPU_ACCESS_FLAG.D3D11_CPU_ACCESS_READ,
                 MiscFlags = 0,
             };
             _device!.CreateTexture2D(in desc, null, out var staging);
